FarmX-Farmer-Assistance-Django/requirements.txt
Django==3.0.4
stripe==2.49.0
requests==2.23.0
djangorestframework==3.11.0
Pillow==7.2.0
